Disable ImportStateVerify for imports in tests
At import time, the YAML/TF manifest is not available. Therefor
the provider can only store the entire API response in manifest.

On the next plan, this will cause a plan with changes because
the read method only stores the attributes set in the YAML/TF
and the values the API returned for them in state. It does not
store all the additional attributes the API returns.

This isn't great, but imports are the exception and the benefits
of the change to not rely on the lastAppliedConfig annotation
anymore overweigh this drawback.