Fix CRD tests
The CRD tests have always added the test-key attribute. But this
attribute was not set in the CRD's spec. And hence, the test-key
attribute has never been persisted by the K8s API.

But because the provider only stored the last config it applied
in the Terraform state, and only diffed against that this bug
never caused the tests to fail.

Now, that we restore the actual API response in the state, this
incorrect CRD spec surfaced in the tests.